Move all types to /types directory, and some constants to /constants directory
