+++ src/rateEngine/RateComponent.ts
@@ -4,19 +4,11 @@ import sum from 'lodash/sum';
 import BillingDeterminants from './billingDeterminants/_BillingDeterminants';
 import { multiplyDecimals } from './utils/decimals';
 import BillingDeterminantFactoryInterface from './BillingDeterminantFactory';
-import { RateElementFilterArgs } from './RateElement';
-
-interface RateComponentArgs {
-  charge: number | Array<number>;
-  name: string;
-  billingDeterminants: BillingDeterminants;
-}
-
-export type RateComponentInterface = BillingDeterminantFactoryInterface &
-  RateElementFilterArgs & {
-  charge: number | Array<number>;
-  name: string;
-};
+import type {
+  RateComponentArgs,
+  RateComponentInterface,
+  RateElementFilterArgs,
+} from './types';
 
 const MONTHS_PER_YEAR = 12;
