+++ src/rateEngine/RateElement.ts
@@ -1,48 +1,19 @@
 import RateComponent from './RateComponent';
 import RateCalculator from './RateCalculator';
-import BillingDeterminantFactory, {
-  BillingDeterminantFactoryInterface,
-  RateElementType,
-} from './BillingDeterminantFactory';
+import BillingDeterminantFactory from './BillingDeterminantFactory';
 import sum from 'lodash/sum';
 import LoadProfile from './LoadProfile';
 import ValidatorFactory from './ValidatorFactory';
-import { RateComponentInterface } from './RateComponent';
-import { Error } from './validators/_Validator';
 import SurchargeAsPercent from './billingDeterminants/SurchargeAsPercent';
 import PriceProfile from './PriceProfile';
-
-export interface RateElementInterface {
-  id?: string;
-  rateElementType: RateElementType;
-  rateComponents?: Array<RateComponentInterface>;
-  name: string;
-  billingCategory?: BillingCategory;
-  priceProfile?: Array<number> | PriceProfile;
-}
-
-export enum RateElementClassification {
-  ENERGY = 'energy',
-  DEMAND = 'demand',
-  FIXED = 'fixed',
-  SURCHARGE = 'surcharge',
-};
-
-// Since the billing category is passed into a RateCalculator as part of the
-// arguments, this enum exists for documentation rather than actual type safety.
-// Once this compiles to JS, nothing will prevent an application from passing in
-// an abitrary string for the billingCategory.
-export enum BillingCategory {
-  TAX = 'tax',
-  SUPPLY = 'supply',
-  DELIVERY = 'delivery',
-}
-
-export interface RateElementFilterArgs {
-  ids?: Array<string>;
-  classifications?: Array<RateElementClassification>;
-  billingCategories?: Array<BillingCategory>;
-}
+import { BillingCategory, RateElementClassification } from './constants';
+import type {
+  RateElementType,
+  RateElementInterface,
+  RateElementFilterArgs,
+  ValidatorError as Error,
+  RateComponentInterface,
+} from './types';
 
 class RateComponentsFactory {
   static make(
