--- src/rateEngine/RateInterface.ts
@@ -1,7 +1,0 @@
-import { RateElementInterface } from './RateElement';
-
-export default interface RateInterface {
-  name: string;
-  title: string;
-  rateElements: Array<RateElementInterface>;
-}
