+++ src/rateEngine/__tests__/LoadProfileScaler.unmocked.test.ts
@@ -2,9 +2,9 @@ import LoadProfileScaler from '../LoadProfileScaler';
 import times from 'lodash/times';
 import sum from 'lodash/sum';
 import LoadProfile from '../LoadProfile';
-//import goalSeek from 'goal-seek';
 import e1 from '../__mocks__/rates/e-1';
-import RateCalculator, { RateCalculatorInterface } from '../RateCalculator';
+import RateCalculator from '../RateCalculator';
+import type { RateCalculatorInterface } from '../types';
 
 const getLoadProfileOfOnes = () => times(8760, () => 1);
