+++ src/rateEngine/__tests__/RateElement.HourlyEnergy.test.ts
@@ -1,7 +1,7 @@
 import PriceProfile from '../PriceProfile';
 import LoadProfile from '../LoadProfile';
 import RateElement from '../RateElement';
-import { RateElementType } from '../BillingDeterminantFactory';
+import { RateElementType } from '../types';
 
 const YEAR = 2019;
 const priceData = Array(8760).fill(0);
