+++ src/rateEngine/billingDeterminants/_BillingDeterminants.ts
@@ -1,13 +1,5 @@
-import { default as lodashMean } from 'lodash/mean';
-import { RateElementClassification } from '../RateElement';
-
-export enum BillingDeterminantsUnits {
-  KWH = 'kWh',
-  KW = 'kW',
-  DAYS = 'days',
-  MONTHS = 'months',
-  DOLLARS = 'dollars',
-}
+import lodashMean from 'lodash/mean';
+import { BillingDeterminantsUnits, RateElementClassification } from '../constants';
 
 abstract class BillingDeterminants {
   abstract rateElementType: string;
